Move to Startup to minimal API
